chore(progress-bar): update progress-bar stories to sb v7
