+++ packages/carbon-web-components/.storybook/main.ts
@@ -26,6 +26,8 @@ const stories = glob.sync(
     '../src/**/file-uploader.stories.ts',
     '../src/**/overflow-menu.mdx',
     '../src/**/overflow-menu.stories.ts',
+    '../src/**/progress-bar.mdx',
+    '../src/**/progress-bar.stories.ts',
   ],
   {
     ignore: ['../src/**/docs/*.mdx'],
